+++ scm-core/src/main/java/sonia/scm/repository/api/LookupCommandBuilder.java
@@ -21,46 +21,31 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-    
+
 package sonia.scm.repository.api;
 
+import sonia.scm.repository.spi.LookupCommand;
+import sonia.scm.repository.spi.LookupCommandRequest;
+
+import java.util.Optional;
+
 /**
- * Enumeration of available commands.
+ * The lookup command executes a lookup for additional repository information.
  *
- * @author Sebastian Sdorra
- * @since 1.17
+ * @since 2.10.0
  */
-public enum Command
-{
-  LOG, BROWSE, CAT, DIFF, BLAME,
+public class LookupCommandBuilder {
 
-  /**
-   * @since 1.18
-   */
-  TAGS,
+  private final LookupCommand lookupCommand;
 
-  /**
-   * @since 1.18
-   */
-  BRANCHES,
+  public LookupCommandBuilder(LookupCommand lookupCommand) {
+    this.lookupCommand = lookupCommand;
+  }
 
-  /**
-   * @since 1.31
-   */
-  INCOMING, OUTGOING, PUSH, PULL,
-  
-  /**
-   * @since 1.43
-   */
-  BUNDLE, UNBUNDLE,
-
-  /**
-   * @since 2.0
-   */
-  MODIFICATIONS, MERGE, DIFF_RESULT, BRANCH, MODIFY,
-
-  /**
-   * @since 2.10.0
-   */
-  LOOKUP;
+  public <T> Optional<T> lookup(Class<T> type, String... args) {
+    LookupCommandRequest<T> request = new LookupCommandRequest<>();
+    request.setType(type);
+    request.setArgs(args);
+    return lookupCommand.lookup(request);
+  }
 }