+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Lookup command which provides further repository information ([#1415](https://github.com/scm-manager/scm-manager/pull/1415))
+
 ### Fixed
 - Error on repository initialization with least-privilege user ([#1414](https://github.com/scm-manager/scm-manager/pull/1414))
 - Adhere to git quiet flag ([#1421](https://github.com/scm-manager/scm-manager/pull/1421))